BackEnd/requirements.txt
brotli==1.0.9
ConfigParser==5.2.0
cryptography==2.8
Cython==0.29.27
Django==3.0
djangorestframework==3.13.1
dl==0.1.0
HTMLParser==0.0.2
ipaddr==2.2.0
keyring==23.5.0
lxml==4.5.0
numpy==1.20.2
ordereddict==1.1
protobuf==3.19.4
PyMySQL==1.0.2
pyOpenSSL==22.0.0
REfO==0.13
spacy==2.3.7
SPARQLWrapper==1.8.5
tqdm==4.57.0
wincertstore==0.2.1
xmlrpclib==1.0.1
